+++ example/polygon_finder_example.dart
@@ -1,46 +1,16 @@
+import 'dart:math';
+
 import 'package:polygon_finder/polygon_finder.dart';
 
-void main() {}
+void main() {
+  List<Polygon> polygons = PolygonFinder.polygonsFromSegments([
+    Line(Point(0, 0), Point(2, 2)),
+    Line(Point(1, 2), Point(1, 0)),
+    Line(Point(-1, 0), Point(1, 0)),
+  ]);
 
-
-
-  // lines.add(Line((0, 0), (10, 10)).toLineString());
-  // lines.add(Line((185, 221), (100, 100)).toLineString());
-  // lines.add(Line((185, 221), (88, 275)).toLineString());
-  // lines.add(Line((88, 275), (180, 316)).toLineString());
-  // lines.add(Line((185, 221), (292, 281)).toLineString());
-  // lines.add(Line((292, 281), (180, 316)).toLineString());
-  // lines.add(Line((189, 98), (83, 187)).toLineString());
-  // lines.add(Line((83, 187), (185, 221)).toLineString());
-  // lines.add(Line((189, 98), (325, 168)).toLineString());
-  // lines.add(Line((325, 168), (185, 221)).toLineString());
-
-  // lines.add(Line((0, 0), (20, 0)).toLineString());
-  // lines.add(Line((0, 10), (20, 10)).toLineString());
-  // lines.add(Line((0, 20), (20, 20)).toLineString());
-  // lines.add(Line((0, 0), (0, 20)).toLineString());
-  // lines.add(Line((10, 0), (10, 20)).toLineString());
-  // lines.add(Line((20, 0), (20, 20)).toLineString());
-
-  // MULTILINESTRING((0 0, 2 2), (1 2, 1 0), (-1 0, 1 0))
-
-  // MULTILINESTRING ((100 100, 200 100, 200 200, 100 200, 100 100),
-  // (120 180, 180 180, 180 120, 120 120, 120 180))
-
-  // Line((0, 0), (10, 0)).toLineString(),
-  // Line((10, 0), (10, 10)).toLineString(),
-  // Line((10, 10), (0, 10)).toLineString(),
-  // Line((0, 10), (0, 0)).toLineString(),
-
-  //
-
-  // List<Polygon> polygons = PolygonFinder.polygonsFromSegments([
-  //   LineSegment(Point(0, 0), Point(2, 2)),
-  //   LineSegment(Point(1, 2), Point(1, 0)),
-  //   LineSegment(Point(-1, 0), Point(1, 0)),
-  // ]);
-
-  // print('Number of polygons: ${polygons.length}');
-  // for (final polygon in polygons) {
-  //   print(polygon);
-  // }
+  print('Number of polygons: ${polygons.length}');
+  for (final polygon in polygons) {
+    print(polygon); // POLYGON((1.0 1.0, 0.0 0.0, 1.0 0.0, 1.0 1.0))
+  }
+}