+++ lib/polygon_finder.dart
@@ -1,5 +1,5 @@
 library polygon_finder;
 
 export 'src/polygon_finder_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.
+export 'src/line.dart';
+export 'src/polygon.dart';